Relocate OS-agnostic Supervisor methods
Move the now OS-agnostic Supervisor methods maybeKillPidFile,
shouldKillProcess and killPid from the UNIX-specific file into the
general one.
