+++ pkg/supervisor/supervisor.go
@@ -18,11 +18,13 @@ package supervisor
 
 import (
 	"context"
+	"errors"
 	"fmt"
 	"os"
 	"os/exec"
 	"path"
 	"runtime"
+	"slices"
 	"sort"
 	"strconv"
 	"strings"
@@ -140,7 +142,11 @@ func (s *Supervisor) Supervise() error {
 	}
 
 	if err := s.maybeKillPidFile(); err != nil {
-		return err
+		if !errors.Is(err, errors.ErrUnsupported) {
+			return err
+		}
+
+		s.log.WithError(err).Warn("Old process cannot be terminated")
 	}
 
 	var ctx context.Context
@@ -240,6 +246,110 @@ func (s *Supervisor) Stop() {
 	if s.done != nil {
 		<-s.done
 	}
+}
+
+// maybeKillPidFile checks kills the process in the pidFile if it's has
+// the same binary as the supervisor's and also checks that the env
+// `_KOS_MANAGED=yes`. This function does not delete the old pidFile as
+// this is done by the caller.
+func (s *Supervisor) maybeKillPidFile() error {
+	pid, err := os.ReadFile(s.PidFile)
+	if os.IsNotExist(err) {
+		return nil
+	} else if err != nil {
+		return fmt.Errorf("failed to read PID file %s: %w", s.PidFile, err)
+	}
+
+	p, err := strconv.Atoi(strings.TrimSuffix(string(pid), "\n"))
+	if err != nil {
+		return fmt.Errorf("failed to parse PID file %s: %w", s.PidFile, err)
+	}
+
+	ph, err := newProcHandle(p)
+	if err != nil {
+		return fmt.Errorf("cannot interact with PID %d from PID file %s: %w", p, s.PidFile, err)
+	}
+
+	if err := s.killProcess(ph); err != nil {
+		return fmt.Errorf("failed to kill PID %d from PID file %s: %w", p, s.PidFile, err)
+	}
+
+	return nil
+}
+
+const exitCheckInterval = 200 * time.Millisecond
+
+// Tries to terminate a process gracefully. If it's still running after
+// s.TimeoutStop, the process is forcibly terminated.
+func (s *Supervisor) killProcess(ph procHandle) error {
+	// Kill the process pid
+	deadlineTicker := time.NewTicker(s.TimeoutStop)
+	defer deadlineTicker.Stop()
+	checkTicker := time.NewTicker(exitCheckInterval)
+	defer checkTicker.Stop()
+
+Loop:
+	for {
+		select {
+		case <-checkTicker.C:
+			shouldKill, err := s.shouldKillProcess(ph)
+			if err != nil {
+				return err
+			}
+			if !shouldKill {
+				return nil
+			}
+
+			err = ph.terminateGracefully()
+			if errors.Is(err, syscall.ESRCH) {
+				return nil
+			} else if err != nil {
+				return fmt.Errorf("failed to terminate gracefully: %w", err)
+			}
+		case <-deadlineTicker.C:
+			break Loop
+		}
+	}
+
+	shouldKill, err := s.shouldKillProcess(ph)
+	if err != nil {
+		return err
+	}
+	if !shouldKill {
+		return nil
+	}
+
+	err = ph.terminateForcibly()
+	if errors.Is(err, syscall.ESRCH) {
+		return nil
+	} else if err != nil {
+		return fmt.Errorf("failed to terminate forcibly: %w", err)
+	}
+	return nil
+}
+
+func (s *Supervisor) shouldKillProcess(ph procHandle) (bool, error) {
+	// only kill process if it has the expected cmd
+	if cmd, err := ph.cmdline(); err != nil {
+		if errors.Is(err, syscall.ESRCH) {
+			return false, nil
+		}
+		return false, err
+	} else if len(cmd) > 0 && cmd[0] != s.BinPath {
+		return false, nil
+	}
+
+	//only kill process if it has the _KOS_MANAGED env set
+	if env, err := ph.environ(); err != nil {
+		if errors.Is(err, syscall.ESRCH) {
+			return false, nil
+		}
+		return false, err
+	} else if !slices.Contains(env, k0sManaged) {
+		return false, nil
+	}
+
+	return true, nil
 }
 
 // Prepare the env for exec:
